% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_base.R
\name{importar_base}
\alias{importar_base}
\title{Importar base de dados}
\usage{
importar_base(arq, col = "standard", re = TRUE)
}
\arguments{
\item{arq}{Um vetor caractere de um elemento, com o caminho absoluto ou relativo do arquivo em formto .csv que será importado. O arquivo deve ter separador ';' e marcador de decimal ','.}

\item{col}{Um vetor caractere de um elemento com os códigos dos tipos das colunas similar ao argumento col_type das funções do pacote readr. Define a tipagem das colunas importadas. Por padrão vale 'cciicnncccccccccciinn' para remedições e 'ccicnnccccccccciinn' para primeira medição}

\item{re}{TRUE/FALSE que indica que o arquivo importado é de uma remedição. Caso contrário, é a primeira medição. Padrão é TRUE para remedição.}
}
\value{
Um tibble com o banco de dados formatado da forma apropriada. Veja os datasets \code{data(medicao)} e \code{data(remedicao)}.
}
\description{
Importar base de dados
}
\details{
O arquivo a ser importado deve obrigatoriamente conter as colunas \strong{bloco}, \strong{cap}, \strong{ind} para o caso de primeira medição. Quando for remedição, acrescentam-se as colunas \strong{ind_ant} e \strong{cap_ant} como obrigatórias. Um exemplo de um banco de dados adequado para medição e remedição pode ser visualizado consultando \code{head(medicao)} e \code{head(remedicao)}
}
