% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juncao_composta.R
\name{juncao_composta}
\alias{juncao_composta}
\title{Junta medições já combinadas com uma nova base de dados}
\usage{
juncao_composta(m_composto, m_inicial, ano_inicial)
}
\arguments{
\item{m_composto}{Um data frame retornado de juntar_medicoes().}

\item{m_inicial}{Um data frame com uma medição anterior para ser combinado}

\item{ano_inicial}{Vetor caractere de um elemento contendo o ano ('YYYY') da medição inicial}
}
\value{
Um tibble com a combinação dos data frames
}
\description{
#' As funções juntar_medicoes() e juncao_composta() são importantes pois as árvores não seguem a mesma ordem nas fichas ao longo dos anos, nem mantém a mesma numeração entre inventários. Isso impede que seja feito um simples 'recortar e colar' das variáveis medidas em cada ano. É necessário que as árvores tenham um código único, como uma âncora, de forma que as informações de um código sejam combinadas única e exclusivamente com as informações do mesmo código entre inventários sucessivos.
Para as árvores com mudança de numeração entre inventários, primeiro a função pareia o mesmo código (coluna 'id') das árvores em dois inventários distintos. Caso não encontre o mesmo código nos dois inventários, então a função busca a âncora no código antigo (coluna 'id_ant').
}
