% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juntar_medicoes.R
\name{juntar_medicoes}
\alias{juntar_medicoes}
\title{Juntar Medições de inventários sucessivos}
\usage{
juntar_medicoes(m_final, m_inicial, ano_inicial)
}
\arguments{
\item{m_final}{Um data frame contendo o banco de dados da medição posterior}

\item{m_inicial}{Um data frame contendo o banco de dados da medição inicial}

\item{ano_inicial}{Um vetor caractere de um elemento indicando o ano ('YYYY') da medição inicial}
}
\value{
Um tibble com as medições realizadas nos dois inventários
}
\description{
As funções juntar_medicoes() e juncao_composta() são importantes pois as árvores não seguem a mesma ordem nas fichas ao longo dos anos, nem mantém a mesma numeração entre inventários. Isso impede que seja feito um simples 'recortar e colar' das variáveis medidas em cada ano. É necessário que as árvores tenham um código único, como uma âncora, de forma que as informações de um código sejam combinadas única e exclusivamente com as informações do mesmo código entre inventários sucessivos.
Para as árvores com mudança de numeração entre inventários, primeiro a função pareia o mesmo código (coluna 'id') das árvores em dois inventários distintos. Caso não encontre o mesmo código nos dois inventários, então a função busca a âncora no código antigo (coluna 'id_ant').
}
