% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizinhas_iguais.R
\name{vizinhas_iguais}
\alias{vizinhas_iguais}
\title{Cálculo de árvores vizinhas da mesma espécie}
\usage{
vizinhas_iguais(alvo, n.vizinhas, bif = FALSE, simplify = TRUE)
}
\arguments{
\item{alvo}{índice da árvore alvo dentro do banco de dados}

\item{n.vizinhas}{vetor numérico de um elemento, com o número de árvores vizinhas consideradas.}

\item{bif}{TRUE/FALSE que indica se as bifurcações serão levadas em conta. Padrão é FALSE}

\item{simplify}{TRUE/FALSE que indica se a função deve retornar apenas o número de vizinhas próximas ou um vetor contendo os id's das vizinhas de mesma espécie}
}
\value{
Vetor numérico de um elemento com o número de vizinhas da mesma espécie, quando simplify = TRUE. Vetor caractere contendo os id's das ávores vizinhas de mesma espécie
}
\description{
Conta o número de árvores da mesma espécie dentre as \code{n} vizinhas mais próximas .
}
\details{
\code{bif} leva em conta também as bifurcações das árvores vizinhas, não somente possíveis bifurcações da árvore alvo
}
