% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizinhos_cell.R
\name{vizinhos_cell}
\alias{vizinhos_cell}
\title{Encontra e calcula a distancia das arvores vizinhas}
\usage{
vizinhos_cell(
  celula,
  camada_inferior = camada_sf,
  camada_superior = dados_sf,
  camada_raster = camada,
  ...
)
}
\arguments{
\item{celula}{namero da calula do da camada raster}

\item{camada_inferior}{camada do tipo sf de onde serao retiradas as calulas}

\item{camada_superior}{camada do tipo sf de onde virao os pontos que se deseja saber a distancia}

\item{camada_raster}{camada do tipo raster que fornece as coordenadas das calulas}

\item{...}{argumentos adicionais para serem passados a funcao st_nn}
}
\value{
retorna o mesmo tipo da funcao st_nn()
}
\description{
Encontra e calcula a distancia das arvores vizinhas, dado um ponto de uma calula raster.
}
