% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_base.R
\name{importar_base}
\alias{importar_base}
\title{Importar base de dados}
\usage{
importar_base(arq, col, re = TRUE)
}
\arguments{
\item{arq}{Um vetor caractere de um elemento, com o caminho absoluto ou relativo do arquivo em formto .csv que será importado. O arquivo deve ter separador ';' e marcador de decimal ','.}

\item{col}{Um vetor caractere de um elemento com os códigos dos tipos das colunas similar ao argumento col_type das funções do pacote readr. Define a tipagem das colunas importadas.}

\item{re}{TRUE/FALSE que indica que o arquivo importado é de uma remedição. Caso contrário, é a primeira medição. Padrão é TRUE para remedição.}
}
\value{
Um tibble
}
\description{
Importar base de dados
}
\examples{
base <- importar_base(arq = 'dados.csv', col = 'ccnnccnnniinc', re = TRUE)
}
